Move 'internal/retry2' into 'internal/retry'.
