+++ internal/retry/state.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package retry2
+package retry
 
 import (
 	"context"
